requirements.txt
gspread
numpy
oauth2client
pandas
psycopg2-binary
python-dotenv
streamlit
